[TASK] Add AbstractFunctionalTestCase
